work on exception handling

+++ pypp_cli/data/cpp_template/pypp/pypp_shutil.h
@@ -11,9 +11,21 @@ namespace shutil {
 inline void rmtree(const PyStr &path) {
     std::error_code ec;
     std::filesystem::path fs_path(path.str());
+    if (!std::filesystem::exists(fs_path)) {
+        throw FileNotFoundError(
+            PyStr("The system cannot find the file specified: ") + path);
+    }
+    if (!std::filesystem::is_directory(fs_path)) {
+        throw NotADirectoryError(PyStr("The directory name is invalid: ") +
+                                 path);
+    }
     std::filesystem::remove_all(fs_path, ec);
     if (ec) {
-        throw FileSystemError("shutil::rmtree failed", fs_path);
+        if (ec == std::errc::permission_denied) {
+            throw PermissionError(PyStr("Permission denied: ") + path);
+        }
+        throw OSError(PyStr("[Error code: ") + std::to_string(ec.value()) +
+                      PyStr("] shutil::rmtree failed: ") + path);
     }
 }
 } // namespace shutil
